requirements.txt
# Core dependencies
aiohttp>=3.8.0
asyncio>=3.4.3
click>=8.0.0
numpy>=1.20.0
pandas>=1.3.0
psutil>=5.9.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
python-dotenv>=0.19.0
pyyaml>=6.0
requests>=2.27.0
sqlalchemy>=1.4.0
tiktoken>=0.3.0

# Database and vector storage
faiss-cpu>=1.7.0
sqlalchemy>=1.4.0
alembic>=1.7.0

# NLP and machine learning
nltk>=3.6.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
spacy>=3.2.0
transformers>=4.18.0

# Monitoring and metrics
prometheus-client>=0.14.0

# Development tools
black>=22.1.0
flake8>=4.0.0
mypy>=0.931
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0